bmt, pot: move into swarm package (#603)
* bmt, pot: moved into swarm package
