+++ swarm/network/discovery.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/ethereum/go-ethereum/pot"
+	"github.com/ethereum/go-ethereum/swarm/pot"
 )
 
 // discovery bzz extension for requesting and relaying node address records
