+++ swarm/network/kademlia.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/log"
-	"github.com/ethereum/go-ethereum/pot"
+	"github.com/ethereum/go-ethereum/swarm/pot"
 )
 
 /*
